refactor: EvaluationServiceから制約チェック関数を抽出
- constraintCheckers.tsを新規作成
- 7つの制約チェック関数を独立した関数として抽出:
  - isBusinessDay: 営業日判定
  - checkStaffShortage: 人員不足チェック
  - checkConsecutiveWorkViolation: 連勤超過チェック
  - checkNightRestViolation: 夜勤後休息チェック
  - checkQualificationMissing: 資格不足チェック
  - checkLeaveRequestIgnored: 休暇希望無視チェック
  - checkTimeSlotPreferenceViolation: 時間帯希望違反チェック
- EvaluationServiceは委譲パターンで関数を呼び出し
- 未カバーだったメソッドのユニットテストを追加（38件全パス）
- evaluationLogic.ts: 1185行→898行（24%削減）
